pyproject.toml
[tool.poetry]
name = "electronics-retail-chain"
version = "0.1.0"
description = ""
authors = ["Denis_L <denis88lyapin@gmail.com>"]
readme = "README.md"
packages = [{include = "electronics_retail_chain"}]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
django = "^5.0"
django-filter = "^23.5"
python-dotenv = "^1.0.0"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
